revert API change
