+++ packages/calypso-analytics/src/utils/debug.js
@@ -6,4 +6,4 @@ import debugFactory from 'debug';
 /**
  * Module variables
  */
-export default debugFactory( 'calypso:analytics:utils' );
+export default debugFactory( 'calypso:analytics' );
